file_cache: fix: add buf/size combo to exact_buf_oracle if file_io rounded size up. trace: split code out of vfs_optimizer (was getting too large) vfs_mount: expose API to unmount all (needed to be able to delete archives after building) vfs_optimizer: safely handle empty connection/node lists; refactored+documented vfs_tree: eliminate 'file in root dir' warning (can happen easily enough)
rest: fixes/minor tweaks

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@3610 3db68df2-c116-0410-a063-a993310a9797

+++ source/simulation/ProductionQueue.cpp
@@ -99,7 +99,7 @@ jsval CProductionQueue::JSI_Get( JSContext* cx, uintN argc, jsval* argv )
 
     int index = ToPrimitive<int>( argv[0] );
 
-	if(index < 0 || index >= m_items.size() )
+	if(index < 0 || index >= (int)m_items.size() )
 	{
 		JS_ReportError( cx, "Production queue index out of bounds: %d", index );
 		return JSVAL_NULL;
@@ -115,7 +115,7 @@ bool CProductionQueue::JSI_Cancel( JSContext* cx, uintN argc, jsval* argv )
 
     int index = ToPrimitive<int>( argv[0] );
 
-	if(index < 0 || index >= m_items.size() )
+	if(index < 0 || index >= (int)m_items.size() )
 	{
 		JS_ReportError( cx, "Production queue index out of bounds: %d", index );
 		return false;
